Make it build on FreeBSD go1.21.9 freebsd/amd64

+++ freediskspacebytes.go
@@ -10,5 +10,6 @@ import (
 func FreeDiskSpaceBytes(path string) (uint64, error) {
 	var stat unix.Statfs_t
 	unix.Statfs(path, &stat)
-	return stat.Bavail * uint64(stat.Bsize), nil
+	// This first uint64 cast is necessary for building on FreeBSD (go version go1.21.9 freebsd/amd64)
+	return uint64(stat.Bavail) * uint64(stat.Bsize), nil
 }
